refactor(promotification): move invalid route exports to separate module
